# internal/interfaces/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Authentication Service API
  description: |
    AuthService is a self-hosted, multi-tenant authentication platform for teams that want provider-grade auth without handing user data, pricing leverage, or service-to-service trust to a third party.

    It provides email/password auth, OAuth2 social login, enterprise SSO with SAML/OIDC, SCIM 2.0 directory sync, machine-to-machine client credentials, magic links, TOTP 2FA, adaptive security and step-up, passkeys/WebAuthn, organization RBAC, refresh-token rotation, tenant-scoped JWTs, JWKS, audit logs, REST, gRPC, and an importable Go JWT validator.

    ## What This Replaces

    Use AuthService when you would otherwise reach for Auth0, Clerk, Firebase Auth, Supabase Auth, Cognito, Stytch, WorkOS, or Keycloak but need a lightweight service your company can own, deploy, inspect, and extend.

    | Need | AuthService Capability |
    |------|------------------------|
    | Product authentication | Signup, login, refresh, logout, profile, password change |
    | Passwordless login | Magic links and passkeys/WebAuthn |
    | MFA | TOTP setup, enable, login verify, and disable lifecycle |
    | Adaptive security | Risk-driven MFA, remembered devices, admin/user step-up, and policy-driven challenge/block/notify decisions |
    | Social identity | Google, GitHub, Microsoft, and Apple OAuth2 |
    | Enterprise SSO | Per-client SAML 2.0 and OIDC connections, domain routing, metadata, and JIT provisioning |
...

...

...
servers:
...
paths:
...
  /api/auth/organizations:
    get:
      operationId: listOrganizations
      summary: List organizations for the current user
      tags: [Organizations]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: User organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationListResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createOrganization
      summary: Create an organization
      description: Creates a B2B organization under the current client. The authenticated user becomes the `owner`.
      tags: [Organizations]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
            example:
              name: Acme Inc
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMembershipDetails"
        "400":
          description: Invalid organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Organization slug already exists for this client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/organizations/{org_id}:
    get:
      operationId: getOrganization
      summary: Get an organization
      tags: [Organizations]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationID"
      responses:
        "200":
          description: Organization and caller membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMembershipDetails"
        "403":
          description: Caller is not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateOrganization
      summary: Update organization profile
      description: Requires `org:write`.
      tags: [Organizations]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "403":
          description: Missing organization write permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/auth/organization-invitations/accept:
    post:
      operationId: acceptOrganizationInvitation
      summary: Accept an organization invitation
      description: The authenticated user's email must match the invitation email.
      tags: [Organizations]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptOrganizationInvitationRequest"
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMembershipDetails"
        "400":
          description: Invalid or expired invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/clients/{id}:
    get:
      operationId: getClient
      summary: Get a client by ID
      description: Returns a single client's details.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateClient
      summary: Update a client
      description: Updates mutable client settings such as allowed origins, webhook URL, status, and tenant WebAuthn policy controls.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
            example:
              settings:
                webauthn_attestation: direct
                webauthn_require_attestation: true
                webauthn_allowed_attestation_formats: [packed, tpm, apple]
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Client API key — identifies the tenant
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token (JWT)
    SCIMBearerAuth:
      type: http
      scheme: bearer
      description: Per-directory SCIM provisioning token
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Break-glass master admin key for bootstrap or emergencies; prefer AdminBearerAuth
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token returned by `/api/admin/auth/login` or `/api/admin/auth/sso`
  parameters:
...
    OrganizationID:
      name: org_id
      in: path
      required: true
      schema:
        type: string
      description: Organization ID
...
  schemas:
...
    CreateOrganizationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 128
          example: Acme Inc
        slug:
          type: string
          maxLength: 80
          description: URL-safe slug. Defaults to a normalized version of name.
          example: acme-inc
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
          example: Acme Corp
        slug:
          type: string
          maxLength: 80
          example: acme-corp
...
    AcceptOrganizationInvitationRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Raw invitation token returned at invite creation.
...
    UpdateClientRequest:
      type: object
      properties:
        name:
          type: string
        allowed_origins:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          format: uri
        status:
          type: string
          enum: [active, suspended]
        settings:
          type: object
          additionalProperties: true
          description: Replaces the client settings object when supplied.

    # ── Responses ─────────────────────────────
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: invalid request
...
    OrganizationRole:
      type: string
      description: Built-in roles are `owner`, `admin`, `member`, and `viewer`. Custom role keys such as `org:billing-admin` are accepted when paired with explicit permissions.
      default: member
      pattern: "^[a-z0-9][a-z0-9:_.-]{1,126}[a-z0-9]$"
      examples:
        - owner
        - org:billing-admin
    OrganizationPermission:
      type: string
      description: Built-in permissions plus custom lower-case `namespace:action` permission keys. Omit a permissions list to use the built-in defaults for built-in roles.
      pattern: "^[a-z0-9][a-z0-9:_.-]{1,126}[a-z0-9]$"
      examples:
        - org:read
        - billing:manage
    Organization:
      type: object
      properties:
        id:
          type: string
        client_id:
          type: string
        name:
          type: string
        slug:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_by_user_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrganizationMembership:
      type: object
      properties:
        id:
          type: string
        client_id:
          type: string
        organization_id:
          type: string
        user_id:
          type: string
        role:
          $ref: "#/components/schemas/OrganizationRole"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationPermission"
        status:
          type: string
          enum: [active]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrganizationMembershipDetails:
      type: object
      properties:
        organization:
          $ref: "#/components/schemas/Organization"
        membership:
          $ref: "#/components/schemas/OrganizationMembership"
...
    OrganizationListResponse:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMembershipDetails"
...
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        allowed_origins:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          nullable: true
          description: Receives signed `audit.event` webhooks when webhook delivery is enabled
        settings:
          type: object
        status:
          type: string
          enum: [active, suspended]
        token_mode:
          type: string
          enum: [v1_hs256, v2_jwks]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...